move user details to utils
