+++ src/src/lib/components/modals/materials/PreviewMaterial.svelte
@@ -6,6 +6,7 @@
 	import { Button, Modal } from 'flowbite-svelte';
 	import { Canvas } from '@threlte/core';
 	import { page } from '$app/stores';
+	import { multipartUploadFile } from '$lib/utils';
 
 	export let open: boolean;
 	export let file: File;
@@ -31,7 +32,7 @@
 		console.log('Upload is being handled');
 
 		if (file.size > 4 * 1024 * 1024) {
-			await multipartUploadFile();
+			await UploadFile();
 		} else {
 			const formData = new FormData();
 
@@ -56,140 +57,37 @@
 		}
 	}
 
-	async function multipartUploadFile() {
+	async function UploadFile() {
 		const toastId = toast.loading('Large File detected, uploading file...');
-		const chunkSize = 5 * 1024 * 1024; // 5MB chunks
-		const chunks = Math.ceil(file.size / chunkSize);
-		const name = generateFileName(file.name);
-
-		// Start the multipart upload
-		const startResponse = await fetch('/api/upload', {
-			method: 'POST',
-			headers: { 'Content-Type': 'application/json' },
-			body: JSON.stringify({
-				action: 'start',
-				filename: name,
-				contentType: file.type
-			})
-		});
-
-		if (!startResponse.ok) {
+		try {
+			const data = await multipartUploadFile(file);
 			toast.dismiss(toastId);
-			toast.error('Failed to start upload');
-			return;
-		}
 
-		const { uploadId } = await startResponse.json();
-		const parts = [];
+			const formData = new FormData();
+			formData.append('link', data.url);
+			formData.append('title', title);
+			formData.append('description', description);
+			formData.append('name', data.name);
 
-		for (let i = 0; i < chunks; i++) {
-			const start = i * chunkSize;
-			const end = Math.min(start + chunkSize, file.size);
-			const chunk = file.slice(start, end);
+			await toast.promise(
+				fetch('?/multiPartUploadFinal', {
+					method: 'POST',
+					body: formData
+				}),
+				{
+					loading: 'Completing upload, please be patient...',
+					success: 'Material uploaded successfully!',
+					error: 'Failed to upload material!'
+				}
+			);
 
-			// Get a signed URL for this part
-			const urlResponse = await fetch('/api/upload', {
-				method: 'POST',
-				headers: { 'Content-Type': 'application/json' },
-				body: JSON.stringify({
-					action: 'getSignedUrl',
-					filename: name,
-					partNumber: i + 1,
-					uploadId
-				})
-			});
-
-			if (!urlResponse.ok) {
-				toast.dismiss(toastId);
-				toast.error('Failed to get signed URL');
-				return;
-			}
-
-			const { signedUrl } = await urlResponse.json();
-
-			// Upload the part using the signed URL (directly to S3)
-			const uploadResponse = await fetch(signedUrl, {
-				method: 'PUT',
-				body: chunk
-			});
-
-			if (!uploadResponse.ok) {
-				toast.dismiss(toastId);
-				toast.error(`Failed to upload part ${i + 1}`);
-				return;
-			}
-
-			parts.push({
-				PartNumber: i + 1,
-				ETag: uploadResponse.headers.get('ETag')
-			});
-		}
-
-		// Complete the multipart upload
-		const completeResponse = await fetch('/api/upload', {
-			method: 'POST',
-			headers: { 'Content-Type': 'application/json' },
-			body: JSON.stringify({
-				action: 'complete',
-				filename: name,
-				uploadId,
-				parts
-			})
-		});
-
-		if (!completeResponse.ok) {
+			const log = 'change at timestamp: ' + new Date().toISOString();
+			mat_change.set(log);
+			document.getElementById('reload_btn_delete')?.click();
+		} catch (e) {
 			toast.dismiss(toastId);
-			toast.error('Failed to complete upload');
-			return;
+			toast.error('Failed to upload file: ' + e);
 		}
-
-		const { fileUrl } = await completeResponse.json();
-
-		toast.dismiss(toastId);
-
-		// Handle the final file URL as needed
-		const formData = new FormData();
-		formData.append('link', fileUrl);
-		formData.append('title', title);
-		formData.append('description', description);
-		formData.append('name', name);
-
-		await toast.promise(
-			fetch('?/multiPartUploadFinal', {
-				method: 'POST',
-				body: formData
-			}),
-			{
-				loading: 'Completing upload, please be patient...',
-				success: 'Material uploaded successfully!',
-				error: 'Failed to upload material!'
-			}
-		);
-
-		const log = 'change at timestamp: ' + new Date().toISOString();
-		mat_change.set(log);
-		document.getElementById('reload_btn_delete')?.click();
-		console.log('File uploaded successfully:', fileUrl);
-	}
-
-	function generateFileName(originalFileName: string): string {
-		const lastDotIndex = originalFileName.lastIndexOf('.');
-		if (lastDotIndex === -1) {
-			throw new Error('Filename does not have an extension');
-		}
-
-		const extension = originalFileName.substring(lastDotIndex);
-		const timestamp = Math.floor(Date.now() / 1000);
-		const timestampHex = timestamp.toString(16).padStart(8, '0');
-		const randomValue = crypto.getRandomValues(new Uint8Array(5));
-		const randomValueHex = Array.from(randomValue)
-			.map((b) => b.toString(16).padStart(2, '0'))
-			.join('');
-		const counter = Math.floor(Math.random() * 0xffffff);
-		const counterHex = counter.toString(16).padStart(6, '0');
-		const objectId = timestampHex + randomValueHex + counterHex;
-
-		return `${objectId}${extension}`;
 	}
 </script>
